+++ CHANGELOG.md
@@ -4,3 +4,4 @@
   early-schema validation with newer versions of donut.system.
 - Update donut.system dependency from club.donutpower/system 0.0.165
   to party.donut/system 0.0.171.
+- Handle non-string error and validation messages in salmon.signal/signal!