+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## v0.9.0 (2023-07-08)
+
 - Fix a bug in salmon.signal/early-validate-conf that caused it to fail
   early-schema validation with newer versions of donut.system.
 - Update donut.system dependency from club.donutpower/system 0.0.165